+++ clarity/src/vm/database/clarity_db.rs
@@ -82,7 +82,7 @@ impl TryFrom<&str> for StoreType {
     fn try_from(value: &str) -> core::result::Result<Self, Self::Error> {
         use self::StoreType::*;
 
-        let hex_value = u8::from_str_radix(value, 10).map_err(|e| e.to_string())?;
+        let hex_value = value.parse::<u8>().map_err(|e| e.to_string())?;
         match hex_value {
             0x00 => Ok(DataMap),
             0x01 => Ok(Variable),
@@ -506,7 +506,7 @@ impl<'a> ClarityDatabase<'a> {
     }
 
     pub fn put_data<T: ClaritySerializable>(&mut self, key: &str, value: &T) -> Result<()> {
-        self.store.put_data(&key, &value.serialize())
+        self.store.put_data(key, &value.serialize())
     }
 
     /// Like `put()`, but returns the serialized byte size of the stored value
@@ -516,7 +516,7 @@ impl<'a> ClarityDatabase<'a> {
         value: &T,
     ) -> Result<u64> {
         let serialized = value.serialize();
-        self.store.put_data(&key, &serialized)?;
+        self.store.put_data(key, &serialized)?;
         Ok(byte_len_of_serialization(&serialized))
     }
 
@@ -568,7 +568,7 @@ impl<'a> ClarityDatabase<'a> {
 
         let size = serialized.len() as u64;
         let hex_serialized = to_hex(serialized.as_slice());
-        self.store.put_data(&key, &hex_serialized)?;
+        self.store.put_data(key, &hex_serialized)?;
 
         Ok(pre_sanitized_size.unwrap_or(size))
     }
@@ -755,16 +755,14 @@ impl<'a> ClarityDatabase<'a> {
         &mut self,
         contract_identifier: &QualifiedContractIdentifier,
     ) -> Result<Option<ContractAnalysis>> {
-        let x_opt = self
-            .store
+        self.store
             .get_metadata(contract_identifier, AnalysisDatabase::storage_key())
             // treat NoSuchContract error thrown by get_metadata as an Option::None --
             //    the analysis will propagate that as a CheckError anyways.
-            .ok();
-        match x_opt.flatten() {
-            None => Ok(None),
-            Some(x) => ContractAnalysis::deserialize(&x).map(|out| Some(out)),
-        }
+            .ok()
+            .flatten()
+            .map(|x| ContractAnalysis::deserialize(&x))
+            .transpose()
     }
 
     pub fn get_contract_size(
@@ -978,7 +976,7 @@ impl<'a> ClarityDatabase<'a> {
 
 // Get block information
 
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     /// Returns the ID of a *Stacks* block, by a *Stacks* block height.
     ///
     /// Fails if `block_height` >= the "currently" under construction Stacks block height.
@@ -1066,7 +1064,7 @@ impl<'a> ClarityDatabase<'a> {
         let query_tip = self.get_index_block_header_hash(current_height.saturating_sub(1))?;
         Ok(self
             .headers_db
-            .get_stacks_height_for_tenure_height(&query_tip, tenure_height.into()))
+            .get_stacks_height_for_tenure_height(&query_tip, tenure_height))
     }
 
     /// Get the last-known burnchain block height.
@@ -1158,7 +1156,7 @@ impl<'a> ClarityDatabase<'a> {
     ///    This is the highest Stacks block in this fork whose consensus hash is known.
     /// 3. Resolve the parent StacksBlockId to its consensus hash
     /// 4. Resolve the consensus hash to the associated SortitionId
-    /// In Epoch 3+:
+    ///    In Epoch 3+:
     /// 1. Get the SortitionId of the current Stacks tip
     fn get_sortition_id_for_stacks_tip(&mut self) -> Result<Option<SortitionId>> {
         if !self
@@ -1276,8 +1274,7 @@ impl<'a> ClarityDatabase<'a> {
                 InterpreterError::Expect(
                     "FATAL: no winning burnchain token spend record for block".into(),
                 )
-            })?
-            .into())
+            })?)
     }
 
     pub fn get_miner_spend_total(&mut self, block_height: u32) -> Result<u128> {
@@ -1294,8 +1291,7 @@ impl<'a> ClarityDatabase<'a> {
                 InterpreterError::Expect(
                     "FATAL: no total burnchain token spend record for block".into(),
                 )
-            })?
-            .into())
+            })?)
     }
 
     pub fn get_block_reward(&mut self, block_height: u32) -> Result<Option<u128>> {
@@ -1316,7 +1312,6 @@ impl<'a> ClarityDatabase<'a> {
         let reward: u128 = self
             .headers_db
             .get_tokens_earned_for_block(&id_bhh, &epoch)
-            .map(|x| x.into())
             .ok_or_else(|| {
                 InterpreterError::Expect("FATAL: matured block has no recorded reward".into())
             })?;
@@ -1337,7 +1332,7 @@ impl<'a> ClarityDatabase<'a> {
 
 // poison-microblock
 
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     pub fn make_microblock_pubkey_height_key(pubkey_hash: &Hash160) -> String {
         format!("microblock-pubkey-hash::{}", pubkey_hash)
     }
@@ -1360,6 +1355,7 @@ impl<'a> ClarityDatabase<'a> {
         self.store.get_cc_special_cases_handler()
     }
 
+    #[allow(clippy::unnecessary_fallible_conversions)]
     pub fn insert_microblock_poison(
         &mut self,
         height: u32,
@@ -1451,11 +1447,11 @@ impl<'a> ClarityDatabase<'a> {
                 if let PrincipalData::Standard(principal_data) = reporter_principal {
                     Ok((principal_data, seq))
                 } else {
-                    return Err(InterpreterError::Expect(
+                    Err(InterpreterError::Expect(
                         "BUG: poison-microblock report principal is not a standard principal"
                             .into(),
                     )
-                    .into());
+                    .into())
                 }
             })
             .transpose()
@@ -1472,7 +1468,7 @@ fn map_no_contract_as_none<T>(res: Result<Option<T>>) -> Result<Option<T>> {
 }
 
 // Variable Functions...
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     pub fn create_variable(
         &mut self,
         contract_identifier: &QualifiedContractIdentifier,
@@ -1605,7 +1601,7 @@ impl<'a> ClarityDatabase<'a> {
 }
 
 // Data Map Functions
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     pub fn create_map(
         &mut self,
         contract_identifier: &QualifiedContractIdentifier,
@@ -1951,7 +1947,7 @@ impl<'a> ClarityDatabase<'a> {
 
 // Asset Functions
 
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     pub fn create_fungible_token(
         &mut self,
         contract_identifier: &QualifiedContractIdentifier,
@@ -2294,19 +2290,13 @@ impl<'a> ClarityDatabase<'a> {
         let key = ClarityDatabase::make_key_for_account_balance(principal);
         debug!("Fetching account balance"; "principal" => %principal.to_string());
         let result = self.get_data(&key)?;
-        Ok(match result {
-            None => STXBalance::zero(),
-            Some(balance) => balance,
-        })
+        Ok(result.unwrap_or_default())
     }
 
     pub fn get_account_nonce(&mut self, principal: &PrincipalData) -> Result<u64> {
         let key = ClarityDatabase::make_key_for_account_nonce(principal);
         let result = self.get_data(&key)?;
-        Ok(match result {
-            None => 0,
-            Some(nonce) => nonce,
-        })
+        Ok(result.unwrap_or_default())
     }
 
     pub fn set_account_nonce(&mut self, principal: &PrincipalData, nonce: u64) -> Result<()> {
@@ -2316,7 +2306,7 @@ impl<'a> ClarityDatabase<'a> {
 }
 
 // access burnchain state
-impl<'a> ClarityDatabase<'a> {
+impl ClarityDatabase<'_> {
     pub fn get_burn_block_height(&self, sortition_id: &SortitionId) -> Option<u32> {
         self.burn_state_db.get_burn_block_height(sortition_id)
     }
@@ -2328,7 +2318,7 @@ impl<'a> ClarityDatabase<'a> {
     }
 
     pub fn get_stacks_epoch_for_block(&self, id_bhh: &StacksBlockId) -> Result<StacksEpochId> {
-        let burn_block = self.get_burnchain_block_height(&id_bhh).ok_or_else(|| {
+        let burn_block = self.get_burnchain_block_height(id_bhh).ok_or_else(|| {
             InterpreterError::Expect(format!(
                 "FATAL: no burnchain block height found for Stacks block {}",
                 id_bhh