+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 ## [Unreleased]
 
 ### Added
+- Add `tenure_timeout_secs` to the miner for determining when a time-based tenure extend should be attempted.
 
 ### Changed
+
+- Nodes will assume that all PoX anchor blocks exist by default, and stall initial block download indefinitely to await their arrival (#5502)
 
 ## [3.1.0.0.1]